+++ lua/entities/base_piano/cl_library_piano.lua
@@ -1,0 +1,5 @@
+module("piano", package.seeall)
+
+function GetLocal()
+    return LocalPlayer():GetPiano()
+end